Add feature CSS Conditional Rules
The CSS Conditional Rules Module Level 3 allows authors to condition rules based on whether particular property declarations are supported in CSS.

With the `@supports` at-rule it is possible to detect the browser support of CSS features directly with CSS.

## Example
An abstract example with the detection of support for `display: flex` (known as CSS Flexible Box Layout Module):

	@supports (display: flex) {
	  div {
	    display: flex;
	  }
	}

## Browser support
This feature is currently supported in Firefox 17 and Opera 12.1 (please see commit for links to references).

All browsers that support the `@media` rule have an implementation of conditioned media queries.

## Additional notes
A test case for `@supports` can be found on [dabblet](http://dabblet.com/gist/3895764).
Another test case for conditioned `@media` is [also available](http://dabblet.com/gist/3896300).

This feature is set to be not shown on the page as it is not wildly supported and further research might be helpful.