Update features-json/datauri.json
Have you considered changing the title to "Data URIs"? That is the correct term. The IETF RFC appears to not care about the URI/URL distinction, but that was in 1998, and modern resources use the correct term. Nicholas Zakas even addresses this issue in [his article](http://www.nczonline.net/blog/2009/10/27/data-uris-explained/).