SVG: + IE warning & fix
IE's SVG resizing seems to be an [ongoing](http://stackoverflow.com/questions/9777143/svg-in-img-element-proportions-not-respected-in-ie9) [issue](http://stackoverflow.com/questions/7711641/why-doesnt-this-svg-graphic-scale-in-ie9-and-10-preview) including [mobile](https://github.com/twitter/twemoji/issues/27).

I don't want to be self-serving by pointing at my own CodePen fork, but the original didn't perform as well in IE10, so I fixed it.  The copy links to others' research, so only 1 link is needed here.