Edge 12 and 13 do not support touch events
Can verify that Edge 12 and 13 do not support touch events. By default, if "touch-action" CSS property is left at default, Edge eats all touch manipulation of the page and sends little to any of it to the browser (except for the first split second, so clicks still work). If "touch-action" is set to "none", the pointer events work, and touches are turned to mouse events, but touches themselves do not respond.