Credential Management API issue in Chrome 51 ~ 56
In Chrome 51 ~ 56, you cannot use credential set at a.d.com in b.d.com. This was fixed in Chrome 57.

Using credentials across different subdomains were not enforced by the spec. The spec used the term MAY but not MUST. But Chrome chose to support this in Chrome 57, so I'd say it was a bug in versions prior to 57. Let's annotate Chrome 51 ~ 56 as partially supported then.