Add token binding
Token binding is a feature to assign a unique ID to a TLS connection. Using a public-private keypair the client proves that only he is the owner of that unique ID. By binding cookies to this unique ID, session hijacking becomes impossible.

Currently only Chrome supports this behind a flag, as far as I know.

This can be tested using the URL https://unbearable-bc.ping-eng.com:3000/open/headers. If that shows the Sec-Token-Binding header, the feature is supported.

Fixes #2083