Fix incorrect data on let in IE11
The `let` data indicates that `let` is fully supported in IE11, but `let` variables are not correctly bound in `for` loops. This seems similar to Note #4, but the variables are in fact scoped to the `for` block and not the function so it seems like a separate issue.

This behavior can be observed with the following:

```js
for (let i = 0; i < 10; i += 1) {
    setTimeout(function () { console.log(i); }, 200);
}
```

In Chrome, this correctly logs the numbers 0-9. In IE, this logs the number 10, 10 times.