web-app-manifest: Updated Chrome and Edge support.
Chrome was listed as partial support; clarified in the footnote that this is due
to missing installation events (starting with 67, fully supported on Chrome OS
but not other platforms).

Edge was listed as full support, but it is also missing the installation events,
so downgraded to partial support to match Chrome.

Test case: https://killer-marmot.appspot.com/web/ (logs installation events in
the page).