Adjust `display: contents` information
This is not intended to be a complete change ready to be accepted, but
as a starting point for discussion and negotiation. This commit message
at the least should be updated before it’s merged!

I’ve removed note 2 from non-Chromium browsers, because it’s wrong: in
Chrome, the content is inaccessible, which is indeed serious, but in
both Firefox and Safari it was merely the accessibility attributes on
the element that were lost—the content was still in the accessibility
tree. I’ve marked them as “yes”, as it doesn’t seem severe enough to me
to warrant calling it only “partial”: most situations that I’ve
encountered where I would like to use `display: contents` are just on a
container <div> that exists purely for reasons of hierarchy, not on a
<ul> or other such semantic element. This still seemed worth adding
another note for.

A detail that I contemplated including but didn’t end up mentioning:
Firefox was *not* in fact completely fixed in 62, only mostly.
https://bugzilla.mozilla.org/show_bug.cgi?id=1500958 tracks fixing it
for the remaining cases (e.g. buttons). It’s not immediately obvious to
me which cases are fixed and which aren’t, but I’m *guessing* that it’s
elements with a special `appearance` that are still affected.

Some more notes on things to consider:

• The known issue says that iOS Safari 10 and 11 render it as `display:
  inline`, while claiming support; but the compatibility table says “no”
  up to 11.2, and “partial” from 11.4 onwards. This needs to be checked,
  and probably a note added to the “no” for 10–11.2 (if it got fixed in
  11.4?). Things claiming support when they don’t actually support it
  are far more dangerous than things that don’t claim support.

• Chrome’s serious bug was fixed on November 20; I’m not sure which
  final release of Chrome that will correspond to.