Replace IE10 / IE11 testing link
The "not being supported" link which is found in the notes about IE10 / IE11 points to codepen. Codepen does not support IE11, therefore this link isn't useful to see what is supported by IE11. This commit replaces that link with one that points to glitch.com (I have copied the test script from codepen to glitch.) Glitch's 'view' feature loads the source as is, and will work in IE11. The glitch version can be used to see support in IE11.