Updates notes for
> In Chrome the API is enabled on Android only

This is misleading and the API is available on desktop too and supports the downlink, effectiveType & rtt values (as point 4 in the notes mentions). You can see the output from the API here here: https://googlechrome.github.io/samples/network-information/. I think it's better to remove this statement.

> Support is also available on Firefox OS.

No longer relevant, so safe to remove.