Adding <model> element to Can I Use
As suggested in #6121, I'm making an attempt to create a page for `<model>`.

There's no `HTML` category? Only `HTML5`? This is not HTML5, it's much later than version 5. No `media` category? That's where I would have put this, along with HTML. I'm guessing this might throw an error, but I'll try anyway.