Add footnote about CSS Nesting requiring a symbol to nest element selectors (in some browsers) (#6840)
* Update css-nesting.json

Adding a footnote to mark which browsers do not support nesting type selectors without a `&` (or other mitigation so that the selector starts with a symbol). 

Firefox did not ship with this limitation. 

Everyone else who supports Nesting has this limitation, except for Safari Technology Preview 179.

* Update css-nesting.json

Adjusting footnote. `&` is not required. `:is()` could be used instead. Anything that makes the selector start with a symbol.

* Update css-nesting.json

Adding link to Chromium issue to footnote.

* Update css-nesting.json

Putting the `&` in code.

* Update css-nesting.json

Firefox on Android should not have any notes. Mistake corrected.