Add a feature for WebAssembly reference types
This is a post-MVP feature extension for WebAssembly. Since its launch,
several such extensions have been added to the standard and shipped in
browsers. See https://github.com/WebAssembly/proposals and
https://webassembly.org/features/

If this example is accepted, I will add additional entries for the
other finished and/or in-progress wasm extensions.