Update webgpu.json with Safari info (#7127)
Adding the fact WebGPU is available in Safari 17.4+ on both macOS and iOS, after flipping the feature flag. 

Updated the footnote to not say "macOS" so it can be used now for both. Also, replacing "experimental feature" with "feature flag". The word "experimental" is used by Chrome, not Safari. We call it a "feature flag" — which comes in three flavors "testable", "preview", and "stable". (Testable = can be tested by flipping the feature flag. Preview = on in Safari Technology Preview. Stable = on by default for all users.)