Update background-clip-text.json
I noticed that Safari 3.1 and 3.2 were marked "unknown", while Safari 3.2 was the first browser will support. So I did some historical digging, talking to folks at Apple. 

The feature was committed to WebKit on March 20, 2008 (three days after Safari 3.1 shipped). Which means it was not in Safari 3.1, so I marked that not supported. And since Safari 3.2 was shipped with all changes in Nov 2008, that's the first this feature appeared, so I marked 3.2 as supported.

https://github.com/WebKit/WebKit/commit/63b6483d723bc56f9b052bd1de9a810e0cdb8b3f