Mark `background-clip: text` as fully supported + correct history of Safari support  (#7204)
* Minor bug with `background-clip: text` shouldn't mark it as only partial support

* Update background-clip-text.json

I noticed that Safari 3.1 and 3.2 were marked "unknown", while Safari 3.2 was the first browser will support. So I did some historical digging, talking to folks at Apple. 

The feature was committed to WebKit on March 20, 2008 (three days after Safari 3.1 shipped). Which means it was not in Safari 3.1, so I marked that not supported. And since Safari 3.2 was shipped with all changes in Nov 2008, that's the first this feature appeared, so I marked 3.2 as supported.

https://github.com/WebKit/WebKit/commit/63b6483d723bc56f9b052bd1de9a810e0cdb8b3f

* Safari unprefixed `background-clip` in 2010

Can I Use marked Safari as unprefixing in Safari 14, in 2020. That struck me as odd, so I looked into it.

Actually, we unprefixed in Safari 5: https://github.com/WebKit/WebKit/commit/740414aaf941cd54946204d86912fede7bb0b103

I wondered where the Safari 14 date came from. It looks like someone noticed in 2020 that Safari was marked as requiring a prefix. They tested and saw that was wrong, that unprefixed worked. So they just changed the current version of Safari. They did not look into the history. https://github.com/Fyrd/caniuse/issues/6106

Safari 14 is definitely wrong.

* Update Safari on iOS background-clip: text

Filling in the earliest data for Safari on iOS.