Update css-text-box-trim.json (#7216)
Safari 18.2 adds support for `text-box`, `text-box-trim`, and `text-box-edge`. This PR also updates the title of the feature — since the `text-box` shorthand is a better way to use this. And the PR updates the description to explain this includes `text-box-trim`, and `text-box-edge`... and describing more accurately what it does. I also removed the reference to `leading-trim`, since 1) the early draft never shipped in customer-facing browsers, and 2) the name change is very old news.