Correct errors in Push API (#7262)
* Update push-api.json

Updating the footnotes about Safari on macOS. "Partial support" should only apply to Safari 16.1-17.7, since after that it's impossible to run Safari 18.0+ on a version of macOS that does not support Web Push. Changing the note requires splitting one of the footnotes (and thus renumbering the iOS footnotes.)

* Update push-api.json

Updating footnote about web push on iOS and iPadOS. "Delivered silently, meaning no sound, vibration, haptics or screen wake" is not a true statement. Notifications can have sound, vibration, and wake the screen — it's in the user's control whether they do or not, exactly like other push notifications on iOS. More info at: https://webkit.org/blog/13878/web-push-for-web-apps-on-ios-and-ipados/