Update css-anchor-positioning for a Safari bug
Adds a link to a reproduction.

The bug is currently fixed in Safari Technology Preview,
so I didn't file a webkit bug.

In the real world, the library I'm maintaining has to do browser
sniffing
because some of its functionality is broken due to this bug with anchor
positioning in Safari:
https://github.com/pomerantsev/accented/blob/4312cae5b4dde15b09806e84b3ced3e18ea997ac/packages/accented/src/utils/supports-anchor-positioning.ts#L1-L8